Adding the ability to revoke tokens
We keep track of a set of "revoke_tokens", which are stored in the DB, and also in the JWT token.

If the revoke token is removed from the DB, and a user tries to use the corrosponding JWT token, then the request will be denied.

Also adding messages to a lot of exceptions.